check cells before print
